// src/charts/composed-chart/types.ts
export type SeriesPaint = string | echarts.graphic.LinearGradient | ImagePatternObject;
// ...
export type BarVariant =
  | "default"
  | "hatched"
  | "duotone"
  | "duotone-reverse"
  | "gradient"
  | "stripped";